// src/redux/tag-types.ts
export enum tagTypes {
    user = "user",
    admin = "admin",
    faq = "faq",
    blog = "blog",
    services = "services",
    availbility = "availbility",
    booking = "booking",
    category = "category",
    review = "review",
    comment = "comment",
    payment = "payment",
    provider = "provider",
    offer= "offer",
    providerReview = "providerReview",
    combo = "combo",
    comboBooking = "comboBooking"
  }
  
  export const tagTypesList = [
    tagTypes.user,
    tagTypes.admin,
    tagTypes.faq,
    tagTypes.blog,
    tagTypes.services,
    tagTypes.availbility,
    tagTypes.booking,
    tagTypes.category,
    tagTypes.review,
    tagTypes.comment,
    tagTypes.payment,
    tagTypes.provider,
    tagTypes.offer,
    tagTypes.providerReview,
    tagTypes.combo,
    tagTypes.comboBooking
  ];
  